+++ R/orbital_speed.R
@@ -70,8 +70,8 @@ compute_orbital_speeds <- function(
 
   if (output_speeds) {
     out <- array(NA, dim = c(dim(spectral_u_component), 2))
-    out[,, 1] <- spectral_u_component
-    out[,, 2] <- spectral_v_component
+    out[, , 1] <- spectral_u_component
+    out[, , 2] <- spectral_v_component
     return(out)
   } else {
     u_rms <- sqrt(2 * resourcecode::fastTrapz(freq, spectral_u_component, 2))