+++ NEWS.md
@@ -1,3 +1,5 @@
+# resourcecode (development version)
+
 # resourcecode 0.2.1
 
 - Move out the data from the package to `{resourcecodedata}` package to be compliant with